requirements.txt
flask
pandas
numpy
mysql-connector-python
openpyxl
werkzeug
